Refactor product and warehouse models

+++ src/esWMS.Application/Functions/WarehouseUnits/Commands/CreateWarehouseUnit/CreateWarehouseUnitCommand.cs
@@ -15,10 +15,10 @@ namespace esWMS.Application.Functions.WarehouseUnits.Commands.CreateWarehouseUni
         public string WarehouseId { get; set; } = null!;
         public string? MediaId { get; set; }
         public string? LocationId { get; set; }
-        public int? TotalWeight { get; set; }
-        public int? TotalLength { get; set; }
-        public int? TotalWidth { get; set; }
-        public int? TotalHeight { get; set; }
+        public decimal? TotalWeight { get; set; }
+        public decimal? TotalLength { get; set; }
+        public decimal? TotalWidth { get; set; }
+        public decimal? TotalHeight { get; set; }
         public bool IsBlocked { get; set; }
         public bool? CanBeStacked { get; set; }
         public string? StackOnId { get; set; }
